editoast: move some ModelV2 impl blocks in submodules
